+++ packages/iti-react-core/src/_util/getGuid.test.ts
@@ -1,0 +1,5 @@
+import { getGuid } from './getGuid'
+
+it('generates a GUID', () => {
+    expect(getGuid()).toHaveLength(36)
+})